requirements.txt
Django>=2.0,<3.0
psycopg2-binary>=2.8
django-cors-headers
djangorestframework
celery==5.0.5
django-environ
gunicorn==19.9.0
psycopg2
pytz
